Re-factored Cube growth/deletion scripts
Modified scripts for growing & deleting scripts to mimic the strategy
pattern.  Before the player control script contained multiple if
statements to determine the side of the cube that was clicked on and
manage cube creation and deletion accordingly.  Now there is a single
script on each cube side with functions to create & delete cubes and
variables to adjust the behaviors.  The player control script only
invokes those functions to create & delete cubes.